Version 5 - Competely Re-defined the file and code structure to make it easier to understand and make further improvements on.
All styling files have been removed for now to avoid inconsistency and unnecessary complexity.

Worked on contact page and added the required forms.

Worked on homepage and added some content.

This update mainly focuses on improving the simplicty of both the structure of the components and codes to make the app easier to work on.