Version 20 - Enabled the update card functionality (button) by creating a new component "UpdateModal.jsx" and by implementing it.
Also created a new styling file for the new specified component called updatemodal.css for a user friendly update section where users can update the question, answer, or the status of the card.

In addition to implementing all these changes, the delete button and flipping of the cards when clicking on them works as well thanks to major upgrades in the file FlashCards.jsx (page) and minor changes in the FlashCardItem.jsx component.