Version 21.1 - Finally fixed and done with the share function!
Now users can select multiple cards by clicking on the bar under each card which turns the borders and the specified bar to green to let the users know that the card(s) is selected.

Clicking on share after selecting multiple cards will send their details over email in JSON format.

Turns out that even though I have written css code for the bar under "checkbox-container" class (in flashcarditem.css), I have forgetten to add the specified class to the FlashCardItem component.

I also added the necessary code (in FlashCardItem.jsx) to see the Card Status and Last Modification Date of the cards on the front side of it.